Clean up lib.rs
