+++ src/allowance.rs
@@ -1,5 +1,6 @@
-use super::storage_types::{AllowanceDataKey, DataKey};
-use super::{Env, Identifier};
+use crate::public_types::Identifier;
+use crate::storage_types::{AllowanceDataKey, DataKey};
+use stellar_contract_sdk::Env;
 
 pub fn read_allowance(e: &Env, from: Identifier, spender: Identifier) -> u64 {
     let key = DataKey::Allowance(AllowanceDataKey { from, spender });
