+++ src/balance.rs
@@ -1,5 +1,6 @@
-use super::storage_types::DataKey;
-use super::{Env, Identifier};
+use crate::public_types::Identifier;
+use crate::storage_types::DataKey;
+use stellar_contract_sdk::Env;
 
 pub fn read_balance(e: &Env, id: Identifier) -> u64 {
     let key = DataKey::Balance(id);
