+++ src/storage_types.rs
@@ -1,4 +1,4 @@
-use crate::Identifier;
+use crate::public_types::Identifier;
 use stellar_contract_sdk::{contracttype, ConversionError, Env, EnvVal, IntoEnvVal, RawVal};
 
 #[derive(Clone)]
